feat(models): move ledger methods from transaction model to ledger lib
